Nueva version con soporte numeros complejos

+++ src/calc/token.rs
@@ -1,6 +1,8 @@
+use num_complex::Complex64;
+
 #[derive(Debug, Clone, PartialEq)]
 pub enum Token {
-    Number(f64),
+    Number(Complex64),
     Op(char),
     LParen,
     RParen,
@@ -8,4 +10,3 @@ pub enum Token {
     Func(String),
     Var(String),
 }
-
